refactor(agents): hoist shared plugins->skills resolver into agents/_skills.py
Blocker 1 (concrete slice): pi_agent imported opencode_agent's private
_plugin_skill_dirs, coupling the two harnesses through an implementation-private
symbol while a comment claimed a shared-module hoist was "out of scope". Move the
resolver (+ _manifest_skill_dirs and its constants) into a new shared
agents/_skills.py; opencode_agent and pi_agent both import it from there, so the
cross-agent private import is gone. opencode_agent re-exports it, so its existing
test import is unchanged.

The FULL nd-JSON-CLI driver hoist (a shared NdJsonCliAgent base for communicate /
kill / reap / crash / cost across pi + opencode) is a larger refactor of two
harnesses including the production OpenCode one; it is left as a tracked follow-up
rather than bolted onto this review-response, and the deferral is now documented
at the _TERM_GRACE_SECONDS comment instead of the stale "out of scope" note.

214 opencode+pi tests, custom lint (396), check and typecheck all pass.
